+++ src/Selenium2Library/locators/elementfinder.py
@@ -1,7 +1,6 @@
 from Selenium2Library import utils
 from robot.api import logger
 from robot.utils import NormalizedDict
-from robot.libraries.BuiltIn import BuiltIn
 
 
 class ElementFinder(object):
@@ -19,11 +18,9 @@ class ElementFinder(object):
             'jquery': self._find_by_sizzle_selector,
             'sizzle': self._find_by_sizzle_selector,
             'tag': self._find_by_tag_name,
-            'scLocator': self._find_by_sc_locator,
             'default': self._find_by_default
         }
         self._strategies = NormalizedDict(initial=strategies, caseless=True, spaceless=True)
-        self._default_strategies = strategies.keys()
 
     def find(self, browser, locator, tag=None):
         assert browser is not None