+++ CHANGES.rst
@@ -124,6 +124,10 @@ continuous integration builds to go green by fixing internal tests.
 - Fixed issue where 'Select Window’ with url strategy fails to locate window
   [laulaz]
 
+- Allow using key attributes (default strategy) when the locator contains
+  a '='. Also make locator prefixes space-insensitive.
+  [ombre42]
+
 1.5
 ---
 - Copy Desired Capabilities before modifying to prevent affecting future